requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
black==22.12.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.4
kombu==5.2.4
mypy-extensions==0.4.3
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
pytz==2022.6
redis==4.4.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
vine==5.0.0
wcwidth==0.2.5
